Hoist chunked object reads into obstore utils
Address PR #1062 review feedback:
- Move the CrIS-local _read_object_chunked helper to
  earth2studio/data/utils.py as obstore_read_chunked and give
  obstore_fetch_to_cache a chunked option, so CrIS granule downloads go
  through the shared save-to-cache util (cache file names unchanged)
- Relocate the chunked-read unit tests next to the other obstore util
  tests in test_data_utils.py and cover the chunked ValueError guard
- Clarify why VIIRS HDF5 decode runs via asyncio.to_thread despite
  _HDF5_LOCK, and why day-listing memoization waits an hour past the
  UTC day boundary
