+++ earth2studio/data/utils.py
@@ -997,6 +997,78 @@ async def obstore_read_range(
     return bytes(data)
 
 
+# Defaults for chunked whole-object reads. 1 MiB chunks with 8 in flight
+# measured ~5x faster and far less variable than a single whole-object GET on
+# tens-of-MB S3 objects (2--8 MiB chunks yield too few concurrent streams).
+# Callers that also parallelize across files should keep files x chunks near
+# their store's connection-pool size.
+_READ_CHUNK_SIZE_BYTES: int = 1024 * 1024
+_READ_CHUNK_CONCURRENCY: int = 8
+
+
+async def obstore_read_chunked(
+    store: AsyncReadableStore,
+    key: str,
+    chunk_size: int = _READ_CHUNK_SIZE_BYTES,
+    max_concurrent: int = _READ_CHUNK_CONCURRENCY,
+) -> bytes:
+    """Reads a whole object, fetching large objects as parallel byte ranges.
+
+    A single whole-object GET runs at single-stream throughput; splitting
+    large objects (e.g. tens-of-MB HDF5 granules) into ``chunk_size`` byte
+    ranges fetched concurrently recovers multi-stream throughput. Objects of
+    at most ``chunk_size`` bytes are read with a single range GET; larger
+    objects are split with at most ``max_concurrent`` ranges in flight, then
+    reassembled in order.
+
+    Parameters
+    ----------
+    store : AsyncReadableStore
+        obspec-conforming store to read from (e.g. an obstore store)
+    key : str
+        Object key (bucket-relative path)
+    chunk_size : int, optional
+        Byte-range size per GET, by default 1 MiB
+    max_concurrent : int, optional
+        Maximum concurrent range GETs for one object, by default 8
+
+    Returns
+    -------
+    bytes
+        The complete object payload
+
+    Raises
+    ------
+    FileNotFoundError
+        If the object does not exist. obstore's ``NotFoundError`` subclasses
+        plain ``Exception``, so it is translated here (mirroring
+        ``obstore_read_range``) to keep callers' retry semantics sane.
+    """
+    try:
+        meta = await store.head_async(key)
+        size = int(meta["size"])
+
+        if size <= chunk_size:
+            return bytes(await store.get_range_async(key, start=0, end=size))
+
+        semaphore = asyncio.Semaphore(max_concurrent)
+
+        async def _fetch_range(start: int) -> bytes:
+            async with semaphore:
+                return bytes(
+                    await store.get_range_async(
+                        key, start=start, end=min(start + chunk_size, size)
+                    )
+                )
+
+        chunks = await asyncio.gather(
+            *(_fetch_range(start) for start in range(0, size, chunk_size))
+        )
+    except (FileNotFoundError, obs.exceptions.NotFoundError):
+        raise FileNotFoundError(f"Object {key} not found in store") from None
+    return b"".join(chunks)
+
+
 async def obstore_fetch_to_cache(
     store: AsyncReadableStore,
     key: str,
@@ -1004,6 +1076,7 @@ async def obstore_fetch_to_cache(
     byte_offset: int = 0,
     byte_length: int | None = None,
     cache_key: str | None = None,
+    chunked: bool = False,
 ) -> str:
     """Fetches a byte range of an object into a local cache file.
 
@@ -1027,20 +1100,29 @@ async def obstore_fetch_to_cache(
         Number of bytes to read, by default None (read to end)
     cache_key : str | None, optional
         Explicit cache file name, by default None (sha256 of key + offset)
+    chunked : bool, optional
+        Fetch the object as concurrent byte-range chunks via
+        :func:`obstore_read_chunked`. Only valid for whole-object fetches
+        (``byte_offset == 0`` and ``byte_length is None``), by default False
 
     Returns
     -------
     str
         Path to the local cache file
     """
+    if chunked and (byte_offset != 0 or byte_length is not None):
+        raise ValueError("chunked fetches only support whole-object reads")
     if cache_key is None:
         cache_key = sha256((key + str(byte_offset)).encode()).hexdigest()
     cache_path = os.path.join(cache_dir, cache_key)
     if Path(cache_path).is_file():
         return cache_path
-    data = await obstore_read_range(
-        store, key, byte_offset=byte_offset, byte_length=byte_length
-    )
+    if chunked:
+        data = await obstore_read_chunked(store, key)
+    else:
+        data = await obstore_read_range(
+            store, key, byte_offset=byte_offset, byte_length=byte_length
+        )
     await asyncio.to_thread(Path(cache_path).write_bytes, data)
     return cache_path
